+++ routes/index.js
@@ -1,6 +1,6 @@
 var express = require('express');
 var router = express.Router();
-const eventRouter = require('./events')
+const eventRouter = require('./cities')
 
 /* GET home page. */
 router.get('/', function(req, res, next) {
